modules/agent/pyproject.toml
[project]
name = "openai-agents-react"
version = "1.0.0"
description = "Agent backend using OpenAI Agents SDK with React frontend."
authors = [
    { name = "William Hubenschmidt", email = "whubenschmidt@gmail.com" },
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
dependencies = [
  "openai-agents>=0.0.3",
  "python-dotenv>=1.0,<2",
  "openai>=2.6.0,<3",
  "fastapi>=0.119.1,<1",
  "uvicorn>=0.38,<1",
  "websockets>=15,<16",
  "pydantic>=2.0,<3",
  "sendgrid>=6.0,<7",
  "google-search-results>=2.4,<3",
]


[project.optional-dependencies]
dev = ["mypy>=1.11.1", "ruff>=0.6.1"]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["agent"]
[tool.setuptools.package-dir]
"agent" = "src/agent"


[tool.setuptools.package-data]
"*" = ["py.typed"]

[tool.ruff]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    # We actually do want to import from typing_extensions
    "UP035",
    # Relax the convention by _not_ requiring documentation for every function parameter.
    "D417",
    "E501",
]
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]
[tool.ruff.lint.pydocstyle]
convention = "google"

[dependency-groups]
dev = [
    "anyio>=4.7.0",
    "mypy>=1.13.0",
    "pytest>=8.3.5",
    "ruff>=0.8.2",
]
